Add support for sign extend to long and word
TODO: andw only supports sign extending of %ax, %bx, %cx, but
the assembler is generating this instruction on the extended 64-bit mode
registers which may look valid, but isn't supported. For example:

andw $0x45, %r8w

is not a valid x86 isntruction even though %r8w is a 16-bit register.